split up gateway test file
